Remove NumericIndex.__new__ (almost) + various methods
