+++ pandas/core/indexes/range.py
@@ -178,6 +178,17 @@ class RangeIndex(NumericIndex):
         result._reset_identity()
         return result
 
+    @classmethod
+    def _validate_dtype(cls, dtype: Dtype | None) -> None:
+        if dtype is None:
+            return
+
+        validation_func, expected = cls._dtype_validation_metadata
+        if not validation_func(dtype):
+            raise ValueError(
+                f"Incorrect `dtype` passed: expected {expected}, received {dtype}"
+            )
+
     # --------------------------------------------------------------------
 
     # error: Return type "Type[NumericIndex]" of "_constructor" incompatible with return
