+++ pandas/tests/indexes/interval/test_constructors.py
@@ -39,8 +39,8 @@ class ConstructorTests:
         params=[
             ([3, 14, 15, 92, 653], np.int64),
             (np.arange(10, dtype="int64"), np.int64),
-            (NumericIndex(range(-10, 11), dtype=np.int64), np.int64),
-            (NumericIndex(range(10, 31), dtype=np.uint64), np.uint64),
+            (NumericIndex(np.arange(-10, 11, dtype=np.int64)), np.int64),
+            (NumericIndex(np.arange(10, 31, dtype=np.uint64)), np.uint64),
             (NumericIndex(np.arange(20, 30, 0.5), dtype=np.float64), np.float64),
             (date_range("20180101", periods=10), "<M8[ns]"),
             (
